requirements.txt
asgiref>=3.6.0
beautifulsoup4>=4.11.2
certifi>=2022.12.7
charset-normalizer>=3.0.1
Django>=4.1.7
django-steamauth>=1.1.2.3
idna>=3.4
python-dotenv>=1.0.0
python-steam-api>=1.0.8
requests>=2.28.2
soupsieve>=2.4
sqlparse>=0.4.3
urllib3>=1.26.14